requirements.txt
flask
replicate
Gunicorn
google-generativeai
openai